% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{latlonconverter}
\alias{latlonconverter}
\title{Converter}
\usage{
latlonconverter(Lon, Lat)
}
\arguments{
\item{Lon}{longitudes to be converted}

\item{Lat}{latitudes to be converted}
}
\description{
Converter
}
